pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "data-bridge"
version = "0.1.0"
description = "High-performance data access library with Rust backend and zero Python byte handling"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Tech Platform Team", email = "tech-platform@shopee.com" }
]
keywords = ["mongodb", "orm", "rust", "async", "performance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pytest-asyncio>=1.3.0",
    "uvloop>=0.21.0",
]

# Console scripts
[project.scripts]
dbtest = "data_bridge.test.cli:main"

# Optional dependencies for different data sources
[project.optional-dependencies]
mongodb = []      # MongoDB support (included by default)
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]
benchmark = [
    "pytest-benchmark>=4.0.0",
    "pytest-benchmark[histogram]>=4.0.0",
]

[tool.maturin]
module-name = "data_bridge"
python-source = "python"
manifest-path = "crates/data-bridge/Cargo.toml"
features = ["mongodb", "postgres"]  # Build with MongoDB and PostgreSQL support

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "benchmark: performance benchmark tests (deselect with '-m \"not benchmark\"')",
]

[dependency-groups]
dev = [
    "beanie>=2.0.1",
    "pytest-cov>=7.0.0",
    "pymongo>=4.0.0",
    "mongoengine>=0.27.0",
    "pytest-benchmark[histogram]>=4.0.0",
    "motor>=3.7.1",
    "gevent>=25.9.1",
    "matplotlib>=3.10.8",
    "tabulate>=0.9.0",
    "python-dotenv>=1.2.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.27.0",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "sqlalchemy[asyncio]>=2.0.0",
]
